fixing errors

+++ flow/cadence/transactions/link.js
@@ -17,17 +17,17 @@ transaction() {
       signer.link<&Art.Collection{Art.CollectionPublic, NonFungibleToken.CollectionPublic, NonFungibleToken.Receiver, MetadataViews.ResolverCollection}>(Art.CollectionPublicPath, target: Art.CollectionStoragePath)
     }
 
-    if signer.borrow<&NonFungibleToken.Collection(from:GoatedGoats.CollectionStoragePath) != nil !signer.getCapability<&GoatedGoats.Collection{GoatedGoats.GoatsCollectionPublic, NonFungibleToken.CollectionPublic, NonFungibleToken.Receiver, MetadataViews.ResolverCollection}>(GoatedGoats.CollectionPublicPath).check() {
+    if signer.borrow<&NonFungibleToken.Collection(from:GoatedGoats.CollectionStoragePath) != nil  && !signer.getCapability<&GoatedGoats.Collection{GoatedGoats.GoatsCollectionPublic, NonFungibleToken.CollectionPublic, NonFungibleToken.Receiver, MetadataViews.ResolverCollection}>(GoatedGoats.CollectionPublicPath).check() {
         signer.unlink(GoatedGoats.CollectionPublicPath)
         signer.link<&GoatedGoats.Collection{GoatedGoats.GoatsCollectionPublic, NonFungibleToken.CollectionPublic, NonFungibleToken.Receiver, MetadataViews.ResolverCollection}>(GoatedGoats.CollectionPublicPath, target: Goat.CollectionStoragePath)
     }
 
-    if signer.borrow<&NonFungibleToken.Collection(from:Flovatar.CollectionStoragePath) != nil !signer.getCapability<&Flovatar.Collection{Flovatar.CollectionPublic, NonFungibleToken.CollectionPublic, NonFungibleToken.Receiver, MetadataViews.ResolverCollection}>(Flovatar.CollectionPublicPath).check() {
+    if signer.borrow<&NonFungibleToken.Collection(from:Flovatar.CollectionStoragePath) != nil && !signer.getCapability<&Flovatar.Collection{Flovatar.CollectionPublic, NonFungibleToken.CollectionPublic, NonFungibleToken.Receiver, MetadataViews.ResolverCollection}>(Flovatar.CollectionPublicPath).check() {
         signer.unlink(Flovatar.CollectionPublicPath)
         signer.link<&Flovatar.Collection{Flovatar.CollectionPublic, NonFungibleToken.CollectionPublic, NonFungibleToken.Receiver, MetadataViews.ResolverCollection}>(Flovatar.CollectionPublicPath, target: Flovatar.CollectionStoragePath)
     }
 
-    if signer.borrow<&NonFungibleToken.Collection(from:FlovatarComponent.CollectionStoragePath) != nil !signer.getCapability<&FlovatarComponent.Collection{FlovatarComponent.CollectionPublic, NonFungibleToken.CollectionPublic, NonFungibleToken.Receiver, MetadataViews.ResolverCollection}>(FlovatarComponent.CollectionPublicPath).check() {
+    if signer.borrow<&NonFungibleToken.Collection(from:FlovatarComponent.CollectionStoragePath) != nil && !signer.getCapability<&FlovatarComponent.Collection{FlovatarComponent.CollectionPublic, NonFungibleToken.CollectionPublic, NonFungibleToken.Receiver, MetadataViews.ResolverCollection}>(FlovatarComponent.CollectionPublicPath).check() {
         signer.unlink(FlovatarComponent.CollectionPublicPath)
         signer.link<&FlovatarComponent.Collection{FlovatarComponent.CollectionPublic, NonFungibleToken.CollectionPublic, NonFungibleToken.Receiver, MetadataViews.ResolverCollection}>(FlovatarComponent.CollectionPublicPath, target: FlovatarComponent.CollectionStoragePath)
     }
